More renames and delete unused classes
