+++ cdcsdk-server/cdcsdk-server-s3/src/main/java/com/yugabyte/cdcsdk/sink/s3/config/S3SinkConnectorConfig.java
@@ -13,7 +13,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.yugabyte.cdcsdk.sink.s3;
+package com.yugabyte.cdcsdk.sink.s3.config;
 
 import static org.apache.kafka.common.config.ConfigDef.Range.atLeast;
 
@@ -49,9 +49,7 @@ import com.amazonaws.regions.RegionUtils;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.SSEAlgorithm;
-import com.yugabyte.cdcsdk.sink.s3.storage.ComposableConfig;
-import com.yugabyte.cdcsdk.sink.s3.storage.config.StorageCommonConfig;
-import com.yugabyte.cdcsdk.sink.s3.storage.config.StorageSinkConnectorConfig;
+import com.yugabyte.cdcsdk.sink.s3.CompressionType;
 
 public class S3SinkConnectorConfig extends StorageSinkConnectorConfig {
