+++ cdcsdk-server/cdcsdk-server-s3/src/main/java/com/yugabyte/cdcsdk/sink/s3/config/StorageSinkConnectorConfig.java
@@ -13,7 +13,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.yugabyte.cdcsdk.sink.s3.storage.config;
+package com.yugabyte.cdcsdk.sink.s3.config;
 
 import java.util.Arrays;
 import java.util.LinkedList;
@@ -25,8 +25,6 @@ import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.common.config.ConfigDef.Importance;
 import org.apache.kafka.common.config.ConfigDef.Type;
 import org.apache.kafka.common.config.ConfigDef.Width;
-
-import com.yugabyte.cdcsdk.sink.s3.storage.ComposableConfig;
 
 public class StorageSinkConnectorConfig extends AbstractConfig implements ComposableConfig {
