+++ cdcsdk-server/cdcsdk-server-s3/src/test/java/com/yugabyte/cdcsdk/sink/s3/S3ConsumerIT.java
@@ -32,6 +32,7 @@ import com.amazonaws.services.s3.model.S3Object;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.inject.Inject;
+import com.yugabyte.cdcsdk.sink.s3.config.S3SinkConnectorConfig;
 import com.yugabyte.cdcsdk.sink.s3.util.S3Utils;
 
 import io.debezium.testing.testcontainers.PostgresTestResourceLifecycleManager;
